+++ packages/plugin-react-oxc/tsdown.config.ts
@@ -2,11 +2,12 @@ import { defineConfig } from 'tsdown'
 
 export default defineConfig({
   entry: 'src/index.ts',
+  fixedExtension: false,
   dts: true,
   copy: [
     {
       from: 'node_modules/@vitejs/react-common/refresh-runtime.js',
-      to: 'dist/refresh-runtime.js',
+      to: 'dist',
     },
   ],
 })