+++ packages/plugin-rsc/examples/browser-mode/src/framework/entry.rsc.tsx
@@ -14,7 +14,7 @@ import buildServerReferences from 'virtual:vite-rsc-browser-mode/build-server-re
 
 export type RscPayload = {
   root: React.ReactNode
-  returnValue?: unknown
+  returnValue?: { ok: boolean; data: unknown }
   formState?: ReactFormState
 }
 
@@ -38,7 +38,7 @@ export function initialize() {
 
 export async function fetchServer(request: Request): Promise<Response> {
   const isAction = request.method === 'POST'
-  let returnValue: unknown | undefined
+  let returnValue: RscPayload['returnValue'] | undefined
   let formState: ReactFormState | undefined
   let temporaryReferences: unknown | undefined
   if (isAction) {
@@ -51,7 +51,12 @@ export async function fetchServer(request: Request): Promise<Response> {
       temporaryReferences = createTemporaryReferenceSet()
       const args = await decodeReply(body, { temporaryReferences })
       const action = await loadServerAction(actionId)
-      returnValue = await action.apply(null, args)
+      try {
+        const data = await action.apply(null, args)
+        returnValue = { ok: true, data }
+      } catch (e) {
+        returnValue = { ok: false, data: e }
+      }
     } else {
       const formData = await request.formData()
       const decodedAction = await decodeAction(formData)
@@ -65,9 +70,9 @@ export async function fetchServer(request: Request): Promise<Response> {
   const rscStream = renderToReadableStream<RscPayload>(rscPayload, rscOptions)
 
   return new Response(rscStream, {
+    status: returnValue?.ok === false ? 500 : undefined,
     headers: {
       'content-type': 'text/x-component;charset=utf-8',
-      vary: 'accept',
     },
   })
 }