+++ packages/plugin-rsc/examples/react-router/cf/entry.ssr.tsx
@@ -1,9 +1,9 @@
-import { generateHTML } from '../react-router-vite/entry.ssr'
+import handler from '../react-router-vite/entry.ssr'
 
 console.log('[debug:cf-ssr-entry]')
 
 export default {
-  fetch(request: Request, env: any) {
-    return generateHTML(request, (request) => env.RSC.fetch(request))
+  async fetch(request: Request, env: any) {
+    return handler(request, await env.RSC.fetch(request))
   },
 }